tests/submission/PhilipTaylor/canvas/2d.imageData.object.readonly.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.imageData.object.readonly</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.imageData.html">imageData</a>.<a href="index.2d.imageData.object.html">object</a>.readonly</h1>
<p class="desc">ImageData objects properties are read-only</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.imageData.type">2d.imageData.type</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("ImageData objects properties are read-only");
_addTest(function(canvas, ctx) {

var imgdata = ctx.getImageData(0, 0, 10, 10);
var w = imgdata.width;
var h = imgdata.height;
var d = imgdata.data;
imgdata.width = 123;
imgdata.height = 123;
imgdata.data = [100,100,100,100];
_assertSame(imgdata.width, w, "imgdata.width", "w");
_assertSame(imgdata.height, h, "imgdata.height", "h");
_assertSame(imgdata.data, d, "imgdata.data", "d");
_assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");
_assertSame(imgdata.data[1], 0, "imgdata.data[\""+(1)+"\"]", "0");
_assertSame(imgdata.data[2], 0, "imgdata.data[\""+(2)+"\"]", "0");
_assertSame(imgdata.data[3], 0, "imgdata.data[\""+(3)+"\"]", "0");


});
</script>

<div id="log"></div>
</body></html>
